refactor(frontend): extract manual viewer into ManualSubtab
Move the Media tab's manual (PDF/Markdown) panel out of MediaTab.vue into
a self-contained ManualSubtab.vue, mirroring ArtworkSubtab. The subtab now
owns its own scroll (flex column filling the content height) so the viewer
keeps its internal scroll and switching subtabs never forces an outer
scrollbar.

Generated-By: PostHog Code
Task-Id: 1d54c00b-3d49-435b-8a16-971389d1382d
